requirements.txt
django
djangorestframework
pymongo
djongo